% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_mean_var.R
\name{boot_mean_var}
\alias{boot_mean_var}
\title{Returns mean and variance of new sample}
\usage{
boot_mean_var(x, d)
}
\arguments{
\item{x}{Numeric vector}

\item{d}{None}
}
\value{
mean and variance of x
}
\description{
\code{boot_mean_var} This is used to compute a mean and variance within a bootstrap function
}
\examples{
boot_mean_var(x= c(1,2,2,2,2,3))
}
