% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fao_aqua.R
\name{clean_fao_aqua}
\alias{clean_fao_aqua}
\title{Assembles and Cleans FAO's Global Fishery and Aquaculture Production Data from scratch}
\usage{
clean_fao_aqua(path_to_zipfile = NA, output_path = NA)
}
\arguments{
\item{path_to_zipfile}{Name of zipped file of FAO data (if in current directory); otherwise, specify path to file}
}
\value{
A merged, tidy dataset.
}
\description{
\code{clean_fao_aqua}
Add on to rebuild_fish which 'Extracts data from FAO zipped file and merges them into "tidy" (i.e., long) format'
}
\examples{
clean_fao_aqua("./raw_data/GlobalProduction_2019.1.0.zip")
}
