% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fmfo_distributions.R
\name{estimate_fmfo_distributions}
\alias{estimate_fmfo_distributions}
\title{Estimates distributions of species and functional group key feed parameters}
\usage{
estimate_fmfo_distributions(data = NA, output_format = "tidy")
}
\arguments{
\item{data}{dataframe}

\item{output_format}{string of either \code{"tidy"} or "NA"}
}
\value{
tibble with bootstrapped data
}
\description{
\code{estimate_fmfo_distribution} #Takes in a dataframe of feed information of aquaculture species (see example table)
Outputs a tidy format of main feed parameters (FM, FO, FCR, etc.) with
Mean, Standard Deviation, Lower bound, and Upper Bound for each country for each taxon and functional group
}
\examples{
estimate_fmfo_distribution(fmfo_data)
estimate_fmfo_distribution(fmfo_data, output_format=NA)
}
