% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Installs and loads packages
Source: https://gist.github.com/stevenworthington/3178163}
\usage{
ipak(pkg)
}
\description{
Installs and loads packages
Source: https://gist.github.com/stevenworthington/3178163
}
\examples{
ipak("ggplot2")
ipak(c("ggplot2", "plyr", "reshape2", "RColorBrewer", "scales", "grid"))
}
